refactor: move ForceModel class to new force_model package
