+++ gmatpyplus/force_model/__init__.py
@@ -1,0 +1,1 @@
+from .force_model import *
